+++ src/data/models/AtraccionAMedida.java
@@ -1,4 +1,4 @@
-package paqueteTP;
+package data.models;
 
 import java.util.ArrayList;
