+++ src/data/models/ClaseTp.java
@@ -1,4 +1,4 @@
-package paqueteTP;
+package data.models;
 
 import java.util.ArrayList;
 import java.util.Scanner;
