+++ src/resources/in/atracciones.in
