+++ src/resources/in/promociones.in
